oops all components!
Refactored map-components out into multiple files with individual components rather than one files with all of them. I kept all the hex-grid components in one file, but all of the more complex components now have their own file
