+++ app/mapper/page.tsx
@@ -1,5 +1,5 @@
 import { Metadata } from "next"
-import Map from "@/lib/components/map-components/map"
+import Map from "@/lib/components/map"
 
 export const metadata: Metadata = {
   title: "Traveller Mapper Tool",
