+++ lib/components/map-components/edit-form.tsx
@@ -1,363 +1,12 @@
-'use client'
-
-import { GasGiant, MilitaryBase, NavalBase, Planet, ScoutBase } from "./symbols"
-import StarSystem from "../util/starsystem"
-import { createGridIDString, deHexify, determineIfSystem, hexify } from "../util/functions"
-import { randomSystem } from "../util/randomSystem"
-import { useState } from "react"
+import { createGridIDString, deHexify, hexify } from "@/lib/util/functions"
+import { randomSystem } from "@/lib/util/randomSystem"
+import StarSystem from "@/lib/util/starsystem"
+import { EmptyParsec, facilityCode, faction, map } from "@/lib/util/types"
+import { faTrash } from "@fortawesome/free-solid-svg-icons"
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome"
-import { faEdit, faMinus, faPlus, faTrash, faX } from "@fortawesome/free-solid-svg-icons"
-import { EmptyParsec, facilityCode, faction, fullRange, map } from "../util/types"
-import MDParse from "./md-parse"
-import Link from "next/link"
+import { useState } from "react"
 
-// Create a single hex (parsec)
-export const Hex = (props: { id: string, screenReader: boolean, possibleSystem?: boolean, style?: string, map: map, setMap: Function }) => {
-  const { id, possibleSystem, screenReader } = props
-  // split id into x,y values
-  const x = Number(id.substring(0, 2))
-  const y = Number(id.substring(2))
-  // If a system might be generated, 50/50 on if one is created
-  let system: StarSystem | EmptyParsec
-  let hasSystem = determineIfSystem()
-  const alreadyMapped = props.map.systems.find(el => el.x === x && el.y === y)
-  if (!alreadyMapped) {
-    let newMap = { ...props.map }
-    //@ts-expect-error
-    const newEntry: StarSystem | emptyParsec = possibleSystem && hasSystem ? randomSystem("", x, y) : new EmptyParsec(x, y)
-    newMap.systems.push(newEntry)
-    props.setMap(newMap)
-    system = newEntry
-  } else {
-    //@ts-expect-error
-    if (!(alreadyMapped.name === undefined)) {
-      // @ts-expect-error
-      system = new StarSystem(alreadyMapped.x, alreadyMapped.y, alreadyMapped.name, alreadyMapped.starport, alreadyMapped.size, alreadyMapped.atmos, alreadyMapped.hydro, alreadyMapped.pop, alreadyMapped.gov, alreadyMapped.law, alreadyMapped.tech, alreadyMapped.travelCode, alreadyMapped.temp, alreadyMapped.factions, alreadyMapped.culture, alreadyMapped.facilities, alreadyMapped.details, alreadyMapped.gasGiant)
-    }
-    else system = alreadyMapped
-  }
-
-  // determine if system POI has water or is asteroid to determine icon
-  let water = system instanceof StarSystem ? system.hydro != 0 : false
-  let asteroid = system instanceof StarSystem ? system.size == 0 : false
-  let basesVerbose = system instanceof StarSystem && system.getBasesArrayVerbose().length > 0 ? system.getBasesArrayVerbose().toString().replaceAll(",", ", ") : "N/A"
-
-  // screen reader "hex"
-  const ScreenReaderHex = () =>
-    <tr className={`border hover:cursor-pointer`}>
-      <td><Link href={`#${createGridIDString(system.x, system.y)}`} onNavigate={() => { if (typeof window !== undefined) window.location.hash = `#${createGridIDString(system.x, system.y)}` }}>{id}</Link></td>
-      <td>{system instanceof StarSystem ? system.name : ""}</td>
-      <td>{system instanceof StarSystem ? system.getUWPSmall() : ""}</td>
-      <td>{system instanceof StarSystem ? String(system.gasGiant) : ""}</td>
-      <td>{system instanceof StarSystem ? basesVerbose : ""}</td>
-    </tr>
-
-
-  // visual hex
-  const VisualHex = () =>
-    <Link
-      className={`hexagon-out bg-black dark:bg-gray-100 relative flex justify-center items-center no-underline`}
-      id={"hex" + props.id}
-      href={`#${createGridIDString(system.x, system.y)}`}
-      onNavigate={() => { if (typeof window !== undefined) window.location.hash = `#${createGridIDString(system.x, system.y)}` }}
-    // onClick={() => { props.setDetails(system); props.setShowDetails(true); }}
-    >
-      <div className={`hexagon-in bg-white dark:bg-gray-800 flex flex-col items-center justify-between hover: cursor-pointer`}>
-        {/* Travel code ring */}
-        <div className={`absolute right-[26px] top-[15px] rounded-full w-[120px] h-[120px] border-2 ${system instanceof StarSystem && system.travelCode == "A" ? "border-amber-300 dark:border-amber-500" : system instanceof StarSystem && system.travelCode == "R" ? "border-red-500 dark:border-red-700" : "border-white dark:border-slate-800"}`} />
-        {/* Content container */}
-        <div className="text-center relative">
-          <p className="text-center m-0 bg-white dark:bg-slate-800">{props.id}</p>
-          {system instanceof StarSystem ? <>
-            {system.gasGiant ? <GasGiant /> : <></>}
-            {system.facilities.includes("N") ? <NavalBase /> : <></>}
-            {system.facilities.includes("M") ? <MilitaryBase /> : <></>}
-            {system.facilities.includes("S") ? <ScoutBase /> : <></>}
-            <p className="text-md font-bold m-0">{system.starport}</p>
-            <Planet water={water} asteroid={asteroid} />
-          </> : <></>}
-        </div>
-        {system instanceof StarSystem ? <>
-          <p className="m-0 relative top-0.5 h-fit truncate font-bold text-center w-[110px] z-20 bg-white dark:bg-slate-800">{system.name}</p>
-          <p className="m-0 text-xs relative  bg-white dark:bg-slate-800 py-1 text-center z-10">{system.getUWPSmall()}</p>
-        </> : <></>}
-      </div>
-    </Link>
-
-
-  return (
-    <>
-      {screenReader ? <ScreenReaderHex /> : <VisualHex />}
-    </>
-  )
-}
-
-// Create a column of 10 hexes (height of a subsector)
-const HexCol = (props: { id: string, start: number, screenReader: boolean, style?: string, possibleSystem?: boolean, map: map, setMap: Function }) => {
-  const { id, start, style } = props
-  // Create an array of Hexes of specified amount
-  // start and stop are used to determine length and for the first 2 digits of the Hex id
-  const arr = []
-  // @ts-ignore
-  for (let i = start; i < start + 10; i++) {
-    const hexId = i < 10 ? "0" + String(i) : String(i)
-    arr.push(<Hex key={id + hexId} id={id + hexId} possibleSystem={props.possibleSystem} screenReader={props.screenReader} map={props.map} setMap={props.setMap} />)
-  }
-  // Map the array out in a div container
-  if (props.screenReader) return (
-    <>
-      {arr.map((el, i) => {
-        return el
-      })}
-    </>
-  )
-  else return (
-    <div className={`${style}`} id={"col" + id + start}>
-      {arr.map((el, i) => {
-        return el
-      })}
-    </div>
-  )
-}
-
-// Create a double column of 10 hexes, second column is offset for use in a larger scale grid
-// id and start determines x,y label of the initial hex
-const HexColDouble = (props: { id: number, start: number, screenReader: boolean, possibleSystem?: boolean, map: map, setMap: Function }) => {
-  const { id, start } = props
-  // parse first 2 digits of hex id for both columns
-  const id1 = id < 10 ? "0" + String(id) : String(id)
-  const id2 = id + 1 < 10 ? "0" + String(id + 1) : String(id + 1)
-  // Create two hex columns, offset the second to create a tight grid
-
-  if (props.screenReader) {
-    return (
-      <>
-        <HexCol id={id1} start={start} possibleSystem={props.possibleSystem} screenReader={true} map={props.map} setMap={props.setMap} />
-        <HexCol id={id2} start={start} possibleSystem={props.possibleSystem} screenReader={true} map={props.map} setMap={props.setMap} />
-      </>
-    )
-  }
-  return (
-    <div className="relative w-[255px]">
-      <HexCol id={id1} start={start} possibleSystem={props.possibleSystem} screenReader={false} map={props.map} setMap={props.setMap} />
-      <HexCol id={id2} start={start} possibleSystem={props.possibleSystem} screenReader={false} map={props.map} setMap={props.setMap} style="absolute top-[75px] left-[127px]" />
-    </div>
-  )
-}
-
-// Create a hex grid, 8 x 10 hexes
-// startX and startY determines the x,y label for the first hex. All other hexes are based on that. Values are truncated to work within sector dimensions.
-export const Subsector = (props: { startX: 1 | 9 | 17 | 25, startY: 1 | 11 | 21 | 31, generateSystems: boolean, screenReader: boolean, sector?: boolean, map: map, setMap: Function }) => {
-  const { startX, startY, generateSystems, sector } = props
-
-  const Map = () => (
-    <div className="flex relative w-fit mx-auto">
-      <HexColDouble id={startX} start={startY} possibleSystem={generateSystems} screenReader={false} map={props.map} setMap={props.setMap} />
-      <HexColDouble id={startX + 2} start={startY} possibleSystem={generateSystems} screenReader={false} map={props.map} setMap={props.setMap} />
-      <HexColDouble id={startX + 4} start={startY} possibleSystem={generateSystems} screenReader={false} map={props.map} setMap={props.setMap} />
-      <HexColDouble id={startX + 6} start={startY} possibleSystem={generateSystems} screenReader={false} map={props.map} setMap={props.setMap} />
-      {/* Border */}
-      <div className={`absolute top-0 left-[.2in] w-full h-full border pointer-events-none`} />
-    </div>
-  )
-
-  const MapForScreenReader = () => (
-    <table className="mx-auto">
-      <thead>
-        <tr>
-          <th>ID</th>
-          <th>Name</th>
-          <th>UWP</th>
-          <th>Gas Giant</th>
-          <th>Bases</th>
-        </tr>
-      </thead>
-      <tbody>
-        <HexColDouble id={startX} start={startY} possibleSystem={generateSystems} screenReader={true} map={props.map} setMap={props.setMap} />
-        <HexColDouble id={startX + 2} start={startY} possibleSystem={generateSystems} screenReader={true} map={props.map} setMap={props.setMap} />
-        <HexColDouble id={startX + 4} start={startY} possibleSystem={generateSystems} screenReader={true} map={props.map} setMap={props.setMap} />
-        <HexColDouble id={startX + 6} start={startY} possibleSystem={generateSystems} screenReader={true} map={props.map} setMap={props.setMap} />
-      </tbody>
-    </table>
-  )
-
-  // screen reader friendly subsectors
-  // do not include local details panel if part of a sector
-  if (props.screenReader && sector) return <MapForScreenReader />
-  else if (props.screenReader && !sector) {
-    return (
-      <>
-        <MapForScreenReader />
-      </>
-    )
-  }
-  // General subsectors
-  // do not include local details panel if part of a sector
-  else if (sector) return (<Map />)
-  else {
-    return (
-      <div className="p-4 relative max-w-screen max-h-screen overflow-scroll">
-        <Zoom>
-          <Map />
-        </Zoom>
-      </div>
-    )
-  }
-}
-
-// Row of 4 subsectors
-const SubsectorRow = (props: { row: 1 | 2 | 3 | 4, generateSystems: boolean, screenReader: boolean, map: map, setMap: Function }) => {
-  const { row, generateSystems } = props
-  //@ts-expect-error
-  const y: 1 | 11 | 21 | 31 = (row - 1) * 10 + 1
-  return (
-    <div className={`flex relative`} id={`row${row}`}>
-      <Subsector startX={1} startY={y} generateSystems={generateSystems} sector={true} screenReader={props.screenReader} map={props.map} setMap={props.setMap} />
-      <Subsector startX={9} startY={y} generateSystems={generateSystems} sector={true} screenReader={props.screenReader} map={props.map} setMap={props.setMap} />
-      <Subsector startX={17} startY={y} generateSystems={generateSystems} sector={true} screenReader={props.screenReader} map={props.map} setMap={props.setMap} />
-      <Subsector startX={25} startY={y} generateSystems={generateSystems} sector={true} screenReader={props.screenReader} map={props.map} setMap={props.setMap} />
-    </div>
-  )
-}
-
-// Create a full sector, 32 x 40 hex grid
-export const Sector = (props: { generateSystems: boolean, screenReader: boolean, map: map, setMap: Function }) => {
-  const { generateSystems } = props
-  return (
-    <div className="p-4 relative max-w-screen max-h-screen overflow-scroll">
-      <Zoom>
-        <SubsectorRow row={1} generateSystems={generateSystems} screenReader={props.screenReader} map={props.map} setMap={props.setMap} />
-        <SubsectorRow row={2} generateSystems={generateSystems} screenReader={props.screenReader} map={props.map} setMap={props.setMap} />
-        <SubsectorRow row={3} generateSystems={generateSystems} screenReader={props.screenReader} map={props.map} setMap={props.setMap} />
-        <SubsectorRow row={4} generateSystems={generateSystems} screenReader={props.screenReader} map={props.map} setMap={props.setMap} />
-      </Zoom>
-    </div>
-  )
-}
-
-// Details panel
-export const DetailsPanel = (props: { system: StarSystem | EmptyParsec, setSystem: Function, setShowDetails: Function, editable: boolean, setMap: Function, map: map }) => {
-  let { system, setSystem, setShowDetails, editable, setMap, map } = props
-  let [editMode, setEditMode] = useState(false)
-  // if (!system) return <></>
-  return (
-    <>
-      <div className="fixed top-0 left-0 w-screen h-screen bg-white dark:bg-slate-800 opacity-75 z-50" />
-      <div className="fixed top-0 left-0 w-screen h-screen flex items-center justify-center z-50">
-        <div className="w-screen md:w-[650px] h-screen md:h-[850px] overflow-scroll opacity-100 bg-slate-100 dark:bg-slate-800 border rounded md:shadow-lg p-4 z-50 scale-100">
-
-          {/* ========== DISPLAY MODE ========== */}
-
-          {!editMode ?
-            <div className="relative">
-              {/* Title Area */}
-              <div className="">
-                <h2 className="text-center w-full font-bold text-xl mb-0">{system instanceof StarSystem ? `${system.getGridID()} ${system.name}` : `${createGridIDString(system.x, system.y)} Empty Parsec`}</h2>
-                {system instanceof StarSystem ? <>
-                  <p className="text-center w-full font-bold text-xl m-0">{system.getUWPSmall()}</p>
-                  <p className="text-center w-full italic m-0">{system.facilities.toString().replaceAll(",", " ")} {system.getTradeCodes().toString().replaceAll(",", " ")} {system.travelCode}</p>
-                </> : <></>}
-                {/* Close panel button */}
-                <Link
-                  className="hover:cursor-pointer hover:bg-slate-300 dark:hover:bg-slate-800 transition-all absolute top-3 right-3 border rounded h-8 w-8 bg-slate-200 dark:bg-slate-700 flex items-center justify-center"
-                  href="#"
-                  onNavigate={() => { if (typeof window !== undefined) window.location.hash = "" }}
-                >
-                  <FontAwesomeIcon icon={faX} />
-                  <p className="absolute scale-0">Close details panel for {createGridIDString(system.x, system.y)}</p>
-                </Link>
-                {editable ? <button onClick={() => setEditMode(true)} className="hover:cursor-pointer hover:bg-slate-300 dark:hover:bg-slate-800 transition-all absolute top-3 left-3 border rounded h-8 w-8 bg-slate-200 dark:bg-slate-700"><FontAwesomeIcon icon={faEdit} /><span className="absolute scale-0">Edit</span></button> : <></>}
-              </div>
-
-              {system instanceof StarSystem ? <>
-                {/* Starport and Trade */}
-                <div className="border-b my-2 pb-2">
-                  <p className="m-0"><span className="font-bold">Starport</span>: {system.getStarportQuality()} (Cr{system.getRandomBerthingCost()}; Fuel {system.getFuelType()})</p>
-                  <p className="m-0"><span className="font-bold">Facilities</span>: {system.getFacilitiesArrayVerbose().length > 0 ? system.getFacilitiesArrayVerbose().toString().replaceAll(",", ", ") : "N/A"}</p>
-                  <p className="m-0"><span className="font-bold">Bases</span>: {system.getBasesArrayVerbose().length > 0 ? system.getBasesArrayVerbose().toString().replaceAll(",", ", ") : "N/A"}</p>
-                  <p className="m-0"><span className="font-bold">Trade Codes</span>: {system.getTradeCodesVerbose().length > 0 ? system.getTradeCodesVerbose().toString().replaceAll(",", ", ") : "N/A"}</p>
-                </div>
-
-                {/* Physical Characteristics */}
-                <div className="border-b my-2 pb-2">
-                  <p className="m-0"><span className="font-bold">Size</span>: {system.getDiameter()}km ({system.getGravity()}G)</p>
-                  <p className="m-0"><span className="font-bold">Atmosphere</span>: {system.getAtmosphereType()}</p>
-                  <p className="m-0"><strong>Temperature</strong>: {system.getTempType()}</p>
-                  <p className="m-0"><span className="font-bold">Hydrographics</span>: {system.getHydroType()}</p>
-                </div>
-
-                {/* Social Characteristics */}
-                <div className="border-b my-2 pb-2">
-                  <p className="m-0"><span className="font-bold">Population</span>: {system.getPopType()}</p>
-                  <p className="m-0"><span className="font-bold">Government</span>: {system.getGovernmentType(system.gov)}</p>
-                  <p className="m-0"><span className="font-bold">Cultural Quirk</span>: {system.getCultureType()}</p>
-                  <p className="m-0"><span className="font-bold">Law</span>: Level {system.law}</p>
-                </div>
-                {/* Factions */}
-                <h3 className="text-center text-xl font-bold">Factions</h3>
-                {system.getFactionArrayVerbose().map((el, i) => {
-                  return (
-                    <div key={`faction${i}`}>
-                      <h4>Faction {i + 1}{el.name ? ` - ${el.name}` : ""}</h4>
-                      <p className="text-sm italic m-0">{el.strength}, {el.gov} Group</p>
-                      <MDParse content={el.details ? el.details : ""} />
-                    </div>
-                  )
-                })}
-                {/* System notes */}
-                <h3 className="text-center text-xl font-bold">System Notes</h3>
-                {system instanceof StarSystem ? <MDParse content={system.details ? system.details : ""} /> : <></>}
-              </> : <></>}
-            </div> :
-            <>
-              {/* ========== EDIT MODE ========== */}
-              <EditForm system={system} setSystem={setSystem} setEditMode={setEditMode} setMap={setMap} map={map} />
-            </>}
-        </div>
-      </div>
-    </>
-  )
-}
-
-const Zoom = (props: { children: React.ReactNode }) => {
-  let [zoom, setZoom] = useState<1 | 2 | 3 | 4>(4)
-
-  const newZoom = (up: boolean) => {
-    let newZoom: 1 | 2 | 3 | 4 = zoom
-    if (up && zoom < 4) newZoom++
-    if (!up && zoom > 1) newZoom--
-    //@ts-expect-error
-    setZoom(newZoom)
-  }
-  return (
-    <>
-      <div className="fixed top-2 right-2 flex flex-col z-50">
-        <button
-          className={`border text-xs flex items-center justify-center ${zoom < 4 ? "bg-white dark:bg-gray-800" : "bg-gray-200 dark:bg-gray-600"} h-[40px] w-[40px] hover:bg-gray-100 dark:hover:bg-gray-600 disabled:hover:bg-gray-200 dark:disabled:bg-gray-600 hover:cursor-pointer disabled:hover:cursor-auto`}
-          onClick={() => newZoom(true)}
-          disabled={zoom === 4}
-        >
-          <FontAwesomeIcon icon={faPlus} />
-          <p className="scale-0 absolute">Increase Zoom</p>
-        </button>
-        <button
-          className={`border text-xs flex items-center justify-center ${zoom > 1 ? "bg-white dark:bg-gray-800" : "bg-gray-200 dark:bg-gray-600"} h-[40px] w-[40px] hover:bg-gray-100 dark:hover:bg-gray-600 disabled:hover:bg-gray-200 dark:disabled:bg-gray-600 hover:cursor-pointer disabled:hover:cursor-auto`}
-          onClick={() => newZoom(false)}
-          disabled={zoom === 1}
-        >
-          <FontAwesomeIcon icon={faMinus} />
-          <p className="scale-0 absolute">Decrease Zoom</p>
-        </button>
-        <p className="border text-center text-[8px] bg-gray-200 dark:bg-gray-700">Zoom {zoom}</p>
-      </div>
-      <div className={`origin-top-left pb-32 md:pb-28`} style={{ transform: `scale(${zoom === 4 ? "1" : zoom === 3 ? ".75" : zoom === 2 ? ".50" : ".25"})` }}>
-        {props.children}
-      </div>
-    </>
-  )
-}
-
-const EditForm = (props: { system: StarSystem | EmptyParsec, setSystem: Function, setEditMode: Function, setMap: Function, map: map }) => {
+export default function EditForm(props: { system: StarSystem | EmptyParsec, setSystem: Function, setEditMode: Function, setMap: Function, map: map }) {
   const { system, setSystem, setEditMode, map, setMap } = props
   let [hasSystem, setHasSystem] = useState(system instanceof StarSystem)
   let [name, setName] = useState(system instanceof StarSystem ? system.name : "")
