+++ lib/components/map-components/symbols.tsx
